Push match_operator pos conversion to the edge
This allows backend implementations to return Option<usize> and only
have the core API glue handle type conversions

+++ artichoke-backend/src/extn/core/regexp/backend/lazy.rs
@@ -158,7 +158,7 @@ impl RegexpType for Lazy {
         &self,
         interp: &mut Artichoke,
         haystack: &[u8],
-    ) -> Result<Option<Int>, Exception> {
+    ) -> Result<Option<usize>, Exception> {
         self.regexp(interp)?
             .inner()
             .match_operator(interp, haystack)
